requirements.txt
asgiref==3.8.1
Brotli==1.1.0
Django==5.1.3
django-environ==0.11.2
djangorestframework
djongo
dnspython==2.7.0
gunicorn==23.0.0
packaging==24.2
psycopg2==2.9.10
pymongo==3.11.4
python-dotenv==1.0.1
pytz==2024.2
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2024.2
whitenoise==6.8.2
